requirements.txt
streamlit
tensorflow
numpy
pillow
scikit-learn
seaborn
matplotlib
gdown
